refactor: use global toast listener for perps startup errors
