+++ ui/components/app/perps/perps-tab.test.tsx
@@ -23,10 +23,6 @@ jest.mock('./perps-toast', () => ({
   PerpsToastProvider: ({ children }: { children: React.ReactNode }) => children,
 }));
 
-jest.mock('./perps-toast/perps-startup-error-toast-listener', () => ({
-  PerpsStartupErrorToastListener: () => null,
-}));
-
 const mockSubmitRequestToBackground =
   submitRequestToBackground as jest.MockedFunction<
     typeof submitRequestToBackground
