+++ ui/components/app/perps/perps-tab.tsx
@@ -21,7 +21,6 @@ import ErrorBoundary from '../error-boundary/error-boundary';
 import { PerpsView } from './perps-view';
 import { PerpsViewStreamBoundary } from './perps-view-stream-boundary';
 import { PerpsToastProvider } from './perps-toast';
-import { PerpsStartupErrorToastListener } from './perps-toast/perps-startup-error-toast-listener';
 
 /**
  * Perps tab content for the account overview.
@@ -78,7 +77,6 @@ export function PerpsTab() {
 
   return (
     <PerpsToastProvider>
-      <PerpsStartupErrorToastListener />
       <ErrorBoundary key="perps">
         <PerpsViewStreamBoundary>
           <PerpsView />
