+++ ui/components/app/toast-listener/toast-listener.test.tsx
@@ -1,5 +1,10 @@
 import React from 'react';
 import { render } from '@testing-library/react';
+import { tEn } from '../../../../test/lib/i18n-helpers';
+import {
+  PERPS_CONFIRMATION_STARTUP_FLOW,
+  PERPS_STARTUP_ERROR_ROUTE_STATE_KEY,
+} from '../../../pages/confirmations/constants/perps';
 import { ToastListener } from './toast-listener';
 
 const mockUseSelector = jest.fn();
@@ -7,9 +12,27 @@ const mockUseSmartTransactionToasts = jest.fn();
 const mockPerpsDepositToast = jest.fn(() => null);
 const mockUsePerpsWithdrawTransactionToasts = jest.fn();
 const mockIsInteractiveUI = jest.fn();
+const mockNavigate = jest.fn();
+const mockToastError = jest.fn();
+const mockTriggerPerpsWithdrawNavigation = jest.fn();
+const mockT = (key: string) => tEn(key);
+let mockLocation = {
+  pathname: '/',
+  search: '',
+  state: null as unknown,
+};
 
 jest.mock('react-redux', () => ({
   useSelector: (selector: unknown) => mockUseSelector(selector),
+}));
+
+jest.mock('react-router-dom', () => ({
+  useLocation: () => mockLocation,
+  useNavigate: () => mockNavigate,
+}));
+
+jest.mock('../../../hooks/useI18nContext', () => ({
+  useI18nContext: () => mockT,
 }));
 
 jest.mock('../../../../shared/lib/selectors/smart-transactions', () => ({
@@ -32,6 +55,19 @@ jest.mock('../perps/perps-deposit-toast', () => ({
   PerpsDepositToast: () => mockPerpsDepositToast(),
 }));
 
+jest.mock('../perps/hooks/usePerpsWithdrawNavigation', () => ({
+  usePerpsWithdrawNavigation: () => ({
+    trigger: mockTriggerPerpsWithdrawNavigation,
+  }),
+}));
+
+jest.mock('../../ui/toast/toast', () => ({
+  toast: {
+    error: (...args: unknown[]) => mockToastError(...args),
+  },
+  ToastContent: () => null,
+}));
+
 jest.mock('./usePerpsWithdrawTransactionToasts', () => ({
   usePerpsWithdrawTransactionToasts: () =>
     mockUsePerpsWithdrawTransactionToasts(),
@@ -40,6 +76,14 @@ jest.mock('./usePerpsWithdrawTransactionToasts', () => ({
 describe('ToastListener', () => {
   beforeEach(() => {
     jest.clearAllMocks();
+    mockLocation = {
+      pathname: '/',
+      search: '',
+      state: null,
+    };
+    mockTriggerPerpsWithdrawNavigation.mockResolvedValue({
+      route: '/confirm-transaction/withdraw-tx-id',
+    });
   });
 
   function renderToastListener({
@@ -87,5 +131,67 @@ describe('ToastListener', () => {
     expect(mockUseSmartTransactionToasts).not.toHaveBeenCalled();
     expect(mockPerpsDepositToast).not.toHaveBeenCalled();
     expect(mockUsePerpsWithdrawTransactionToasts).not.toHaveBeenCalled();
+  });
+
+  it('shows withdraw startup error toast from route state and clears route state', () => {
+    mockLocation = {
+      pathname: '/',
+      search: '?tab=perps',
+      state: {
+        preserved: 'state',
+        [PERPS_STARTUP_ERROR_ROUTE_STATE_KEY]:
+          PERPS_CONFIRMATION_STARTUP_FLOW.WITHDRAW,
+      },
+    };
+
+    renderToastListener({
+      transactionToastEnabled: false,
+      isInteractive: true,
+    });
+
+    const [toastContent, options] = mockToastError.mock.calls[0];
+
+    expect(toastContent.props.title).toBe(tEn('perpsWithdrawStartErrorTitle'));
+    expect(toastContent.props.description).toBe(
+      tEn('perpsWithdrawStartErrorDescription'),
+    );
+    expect(toastContent.props.actionText).toBe(tEn('tryAgain'));
+    expect(options).toStrictEqual({ id: 'perps-startup-error' });
+    expect(mockNavigate).toHaveBeenCalledWith('/?tab=perps', {
+      replace: true,
+      state: { preserved: 'state' },
+    });
+
+    toastContent.props.onActionClick();
+
+    expect(mockTriggerPerpsWithdrawNavigation).toHaveBeenCalledTimes(1);
+  });
+
+  it('shows deposit startup error toast from route state and clears route state', () => {
+    mockLocation = {
+      pathname: '/perps/trade/BTC',
+      search: '',
+      state: {
+        [PERPS_STARTUP_ERROR_ROUTE_STATE_KEY]:
+          PERPS_CONFIRMATION_STARTUP_FLOW.DEPOSIT,
+      },
+    };
+
+    renderToastListener({
+      transactionToastEnabled: false,
+      isInteractive: true,
+    });
+
+    const [toastContent] = mockToastError.mock.calls[0];
+
+    expect(toastContent.props.title).toBe(tEn('perpsDepositToastErrorTitle'));
+    expect(toastContent.props.description).toBe(
+      tEn('perpsDepositToastErrorDescription'),
+    );
+    expect(toastContent.props.actionText).toBeUndefined();
+    expect(mockNavigate).toHaveBeenCalledWith('/perps/trade/BTC', {
+      replace: true,
+      state: undefined,
+    });
   });
 });
